Refactored Backend words api
