+++ Backend/Controllers/WordController.cs
@@ -21,12 +21,14 @@ namespace BackendFramework.Controllers
     public class WordController : Controller
     {
         public readonly IWordService _wordService;
+        public readonly IWordRepository _wordRepo;
         public readonly ILexiconMerger<LiftObject, LiftEntry, LiftSense, LiftExample> _merger;
 
-        public WordController(IWordService wordService)
+        public WordController(IWordService wordService, IWordRepository repo, ILexiconMerger<LiftObject, LiftEntry, LiftSense, LiftExample> merger)
         {
             _wordService = wordService;
-            _merger = (ILexiconMerger<LiftObject, LiftEntry, LiftSense, LiftExample>)wordService;
+            _merger = merger;
+            _wordRepo = repo;
         }
 
         [EnableCors("AllowAll")]
@@ -36,7 +38,7 @@ namespace BackendFramework.Controllers
         [HttpGet]
         public async Task<IActionResult> Get()
         {
-            return new ObjectResult(await _wordService.GetAllWords());
+            return new ObjectResult(await _wordRepo.GetAllWords());
         }
 
         // DELETE v1/Project/Words
@@ -46,7 +48,7 @@ namespace BackendFramework.Controllers
         public async Task<IActionResult> Delete()
         {
 #if DEBUG
-            return new ObjectResult(await _wordService.DeleteAllWords());
+            return new ObjectResult(await _wordRepo.DeleteAllWords());
 #else
             return new UnauthorizedResult();
 #endif
@@ -60,7 +62,7 @@ namespace BackendFramework.Controllers
             List<string> Ids = new List<string>();
             Ids.Add(Id);
 
-            var word = await _wordService.GetWords(Ids);
+            var word = await _wordRepo.GetWords(Ids);
             if (word.Count == 0)
             {
                 return new NotFoundResult();
@@ -73,7 +75,7 @@ namespace BackendFramework.Controllers
         [HttpPost]
         public async Task<IActionResult> Post([FromBody]Word word)
         {
-            await _wordService.Create(word);
+            await _wordRepo.Create(word);
             return new OkObjectResult(word.Id);
         }
 
@@ -84,7 +86,7 @@ namespace BackendFramework.Controllers
         {
             List<string> ids = new List<string>();
             ids.Add(Id);
-            var document = await _wordService.GetWords(ids);
+            var document = await _wordRepo.GetWords(ids);
             if (document.Count == 0)
             {
                 return new NotFoundResult();
@@ -117,7 +119,7 @@ namespace BackendFramework.Controllers
                 ids.Add(childId);
             }
             ids.Add(mergeWords.parent);
-            var document = await _wordService.GetWords(ids);
+            var document = await _wordRepo.GetWords(ids);
             if (document.Count != ids.Count)
             {
                 return new NotFoundResult();
