+++ Backend/Interfaces/IWordService.cs
@@ -10,18 +10,8 @@ namespace BackendFramework.Interfaces
 {
     public interface IWordService
     {
-        //calls for collection
-        Task<List<Word>> GetAllWords();
-        //called on collection, returns a list of words
-        Task<List<Word>> GetWords(List<string> Ids);
-        Task<Word> Create(Word word);
         Task<bool> Update(string Id, Word word);
         Task<bool> Delete(string Id);
-        Task<bool> DeleteAllWords();
         Task<Word> Merge(MergeWords mergeWords);
-
-        Task<List<Word>> GetFrontier();
-        Task<Word> AddFrontier(Word word);
-        Task<bool> DeleteFrontier(string id);
     }
 }
